+++ src/client.rs
@@ -49,7 +49,7 @@ impl Plugin for ClientPlugin {
                         ClientSet::Reset.run_if(client_just_disconnected),
                     ),
                     ClientSet::Receive,
-                    ClientSet::SyncHierarchy,
+                    (ClientSet::Diagnostics, ClientSet::SyncHierarchy),
                 )
                     .chain(),
             )
@@ -180,7 +180,7 @@ fn apply_change_message(
     params: &mut ReceiveParams,
     message: &[u8],
 ) -> bincode::Result<()> {
-    let end_pos: u64 = message.len().try_into().unwrap();
+    let end_pos = message.len();
     let mut cursor = Cursor::new(message);
     if let Some(stats) = &mut params.stats {
         stats.messages += 1;
@@ -209,7 +209,7 @@ fn apply_change_message(
                     apply_entity_mapping(world, params, cursor)
                 })?;
                 if let Some(stats) = &mut params.stats {
-                    stats.mappings += len as u32;
+                    stats.mappings += len;
                 }
             }
             ChangeMessageFlags::DESPAWNS => {
@@ -217,7 +217,7 @@ fn apply_change_message(
                     apply_despawn(world, params, cursor, message_tick)
                 })?;
                 if let Some(stats) = &mut params.stats {
-                    stats.despawns += len as u32;
+                    stats.despawns += len;
                 }
             }
             ChangeMessageFlags::REMOVALS => {
@@ -225,7 +225,7 @@ fn apply_change_message(
                     apply_removals(world, params, cursor, message_tick)
                 })?;
                 if let Some(stats) = &mut params.stats {
-                    stats.entities_changed += len as u32;
+                    stats.entities_changed += len;
                 }
             }
             ChangeMessageFlags::CHANGES => {
@@ -234,7 +234,7 @@ fn apply_change_message(
                     apply_changes(world, params, cursor, message_tick)
                 })?;
                 if let Some(stats) = &mut params.stats {
-                    stats.entities_changed += len as u32;
+                    stats.entities_changed += len;
                 }
             }
             _ => unreachable!("iteration should yield only named flags"),
@@ -254,7 +254,7 @@ fn buffer_mutate_message(
     buffered_mutations: &mut BufferedMutations,
     message: Bytes,
 ) -> bincode::Result<u16> {
-    let end_pos: u64 = message.len().try_into().unwrap();
+    let end_pos = message.len();
     let mut cursor = Cursor::new(&*message);
     if let Some(stats) = &mut params.stats {
         stats.messages += 1;
@@ -300,7 +300,7 @@ fn apply_mutate_messages(
         match len {
             Ok(len) => {
                 if let Some(stats) = &mut params.stats {
-                    stats.entities_changed += len as u32;
+                    stats.entities_changed += len;
                 }
             }
             Err(e) => result = Err(e),
@@ -397,7 +397,7 @@ fn apply_removals(
     })?;
 
     if let Some(stats) = &mut params.stats {
-        stats.components_changed += len as u32;
+        stats.components_changed += len;
     }
 
     params.queue.apply(world);
@@ -452,7 +452,7 @@ fn apply_changes(
     })?;
 
     if let Some(stats) = &mut params.stats {
-        stats.components_changed += len as u32;
+        stats.components_changed += len;
     }
 
     params.queue.apply(world);
@@ -636,6 +636,12 @@ pub enum ClientSet {
     ///
     /// Runs in [`PreUpdate`].
     Receive,
+    /// Systems that populate Bevy's [`Diagnostics`](bevy::diagnostic::Diagnostics).
+    ///
+    /// Used by `bevy_replicon`.
+    ///
+    /// Runs in [`PreUpdate`].
+    Diagnostics,
     /// Systems that synchronize hierarchy changes in [`ParentSync`](super::parent_sync::ParentSync).
     ///
     /// Used by `bevy_replicon`.
@@ -727,15 +733,15 @@ pub(super) struct BufferedMutate {
 #[derive(Default, Resource, Debug)]
 pub struct ClientReplicationStats {
     /// Incremented per entity that changes.
-    pub entities_changed: u32,
+    pub entities_changed: usize,
     /// Incremented for every component that changes.
-    pub components_changed: u32,
+    pub components_changed: usize,
     /// Incremented per client mapping added.
-    pub mappings: u32,
+    pub mappings: usize,
     /// Incremented per entity despawn.
-    pub despawns: u32,
+    pub despawns: usize,
     /// Replication messages received.
-    pub messages: u32,
+    pub messages: usize,
     /// Replication bytes received in message payloads (without internal messaging plugin data).
-    pub bytes: u64,
+    pub bytes: usize,
 }